refactor: 엔드 포인트 분리
