+++ src/main/java/aws/retrospective/service/UserService.java
@@ -1,7 +1,5 @@
 package aws.retrospective.service;
 
-import aws.retrospective.dto.GetPreSignedUrlRequestDto;
-import aws.retrospective.dto.GetPreSignedUrlResponseDto;
 import aws.retrospective.dto.UpdateUserProfileRequestDto;
 import aws.retrospective.dto.UpdateUserProfileResponseDto;
 import aws.retrospective.entity.User;
@@ -15,14 +13,7 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class UserService {
 
-    private final AmazonS3Service s3Service;
     private final UserRepository userRepository;
-
-    @Transactional
-    public GetPreSignedUrlResponseDto getPreSignedUrl(GetPreSignedUrlRequestDto request) {
-        String preSignedUrl = s3Service.getPresignUrl(request.getFilename());
-        return new GetPreSignedUrlResponseDto(preSignedUrl);
-    }
 
     @Transactional
     public UpdateUserProfileResponseDto updateProfile(Long userId, UpdateUserProfileRequestDto request) {
